move ConfigruationStorage to core
