+++ etomica-modules/src/main/java/etomica/modules/glass/CorrelationSelf2.java
@@ -7,12 +7,12 @@ package etomica.modules.glass;
 import etomica.atom.AtomType;
 import etomica.atom.IAtom;
 import etomica.atom.IAtomList;
+import etomica.data.ConfigurationStorage;
 import etomica.data.DataSourceIndependentSimple;
 import etomica.data.DataTag;
 import etomica.data.IDataSource;
 import etomica.data.types.DataDoubleArray;
 import etomica.data.types.DataFunction;
-import etomica.integrator.Integrator;
 import etomica.space.Space;
 import etomica.space.Vector;
 import etomica.units.dimensions.Length;
